requirements.txt
# Serveur et Framework
fastapi
uvicorn[standard]

# Base de données (Asynchrone + Drivers)
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic

# Templates et Formulaires
jinja2
python-multipart

# Sécurité et Authentification
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous
cryptography
python-dotenv

# Génération PDF
weasyprint